refactoring services for modularity
